ingestion/requirements.txt
requests==2.33.1
python-dotenv==1.2.2
flask==3.1.3
flask-cors==6.0.2
pillow==12.2.0
